Move goal ordering tests to core
