+++ tests/test_goals.py
@@ -16,7 +16,7 @@ from kanren.goals import (
     membero,
     rembero,
 )
-from kanren.core import run, eq, goaleval, lall, lallgreedy, EarlyGoalError
+from kanren.core import eq, goaleval, run
 
 x, y, z, w = var("x"), var("y"), var("z"), var("w")
 
@@ -223,39 +223,6 @@ def test_appendo2():
         results = run(2, (x, y, z, w), appendo(x, y, w), appendo(w, z, t))
         for xi, yi, zi, wi in results:
             assert xi + yi + zi == t
-
-
-def test_goal_ordering():
-    # Regression test for https://github.com/logpy/logpy/issues/58
-
-    def lefto(q, p, lst):
-        if isvar(lst):
-            raise EarlyGoalError()
-        return membero((q, p), zip(lst, lst[1:]))
-
-    vals = var()
-
-    # Verify the solution can be computed when we specify the execution
-    # ordering.
-    rules_greedy = (
-        lallgreedy,
-        (eq, (var(), var()), vals),
-        (lefto, "green", "white", vals),
-    )
-
-    (solution,) = run(1, vals, rules_greedy)
-    assert solution == ("green", "white")
-
-    # Verify that attempting to compute the "safe" order does not itself cause
-    # the evaluation to fail.
-    rules_greedy = (
-        lall,
-        (eq, (var(), var()), vals),
-        (lefto, "green", "white", vals),
-    )
-
-    (solution,) = run(1, vals, rules_greedy)
-    assert solution == ("green", "white")
 
 
 def test_rembero():
